+++ Source/CanvasTimeline.cpp
@@ -45,8 +45,8 @@ void CanvasTimeline::Render()
    ofPushMatrix();
    ofTranslate(mX, mY);
 
-   float startX = ofMap(mCanvas->mLoopStart, mCanvas->mViewStart, mCanvas->mViewEnd, 0, mWidth);
-   float endX = ofMap(mCanvas->mLoopEnd, mCanvas->mViewStart, mCanvas->mViewEnd, 0, mWidth);
+   double startX = ofMap(mCanvas->mLoopStart, mCanvas->mViewStart, mCanvas->mViewEnd, 0, mWidth);
+   double endX = ofMap(mCanvas->mLoopEnd, mCanvas->mViewStart, mCanvas->mViewEnd, 0, mWidth);
 
    if (mClick && (mHoverMode == HoverMode::kStart || mHoverMode == HoverMode::kMiddle))
       startX += mDragOffset.x;
@@ -58,10 +58,10 @@ void CanvasTimeline::Render()
    {
       ofSetColor(150, 150, 150);
       ofNoFill();
-      float quantizedStart = GetQuantizedForX(startX, HoverMode::kMiddle);
-      float quantizedStartX = ofMap(quantizedStart, mCanvas->mViewStart, mCanvas->mViewEnd, 0, mWidth);
-      float quantizedEnd = GetQuantizedForX(endX, HoverMode::kMiddle);
-      float quantizedEndX = ofMap(quantizedEnd, mCanvas->mViewStart, mCanvas->mViewEnd, 0, mWidth);
+      double quantizedStart = GetQuantizedForX(startX, HoverMode::kMiddle);
+      double quantizedStartX = ofMap(quantizedStart, mCanvas->mViewStart, mCanvas->mViewEnd, 0, mWidth);
+      double quantizedEnd = GetQuantizedForX(endX, HoverMode::kMiddle);
+      double quantizedEndX = ofMap(quantizedEnd, mCanvas->mViewStart, mCanvas->mViewEnd, 0, mWidth);
       ofRect(quantizedStartX, 0, quantizedEndX - quantizedStartX, mHeight / 2, 0);
    }
    if (mHoverMode == HoverMode::kMiddle)
@@ -75,8 +75,8 @@ void CanvasTimeline::Render()
    {
       ofSetColor(150, 150, 150);
       ofNoFill();
-      float quantized = GetQuantizedForX(startX, HoverMode::kStart);
-      float quantizedX = ofMap(quantized, mCanvas->mViewStart, mCanvas->mViewEnd, 0, mWidth);
+      double quantized = GetQuantizedForX(startX, HoverMode::kStart);
+      double quantizedX = ofMap(quantized, mCanvas->mViewStart, mCanvas->mViewEnd, 0, mWidth);
       DrawTriangle(quantizedX, 1);
    }
    if (mHoverMode == HoverMode::kStart)
@@ -91,8 +91,8 @@ void CanvasTimeline::Render()
    {
       ofSetColor(150, 150, 150);
       ofNoFill();
-      float quantized = GetQuantizedForX(endX, HoverMode::kEnd);
-      float quantizedX = ofMap(quantized, mCanvas->mViewStart, mCanvas->mViewEnd, 0, mWidth);
+      double quantized = GetQuantizedForX(endX, HoverMode::kEnd);
+      double quantizedX = ofMap(quantized, mCanvas->mViewStart, mCanvas->mViewEnd, 0, mWidth);
       DrawTriangle(quantizedX, -1);
    }
    if (mHoverMode == HoverMode::kEnd)
@@ -106,7 +106,7 @@ void CanvasTimeline::Render()
    ofPopMatrix();
 }
 
-void CanvasTimeline::DrawTriangle(float posX, int direction)
+void CanvasTimeline::DrawTriangle(double posX, int direction)
 {
    ofBeginShape();
    ofVertex(posX, 0);
@@ -116,9 +116,9 @@ void CanvasTimeline::DrawTriangle(float posX, int direction)
    ofEndShape();
 }
 
-float CanvasTimeline::GetQuantizedForX(float posX, HoverMode clampSide)
+double CanvasTimeline::GetQuantizedForX(double posX, HoverMode clampSide)
 {
-   float pos = ((posX / mWidth) * (mCanvas->mViewEnd - mCanvas->mViewStart)) + mCanvas->mViewStart;
+   double pos = ((posX / mWidth) * (mCanvas->mViewEnd - mCanvas->mViewStart)) + mCanvas->mViewStart;
    int measure = CLAMP(int(pos + .5f), 0, mCanvas->GetLength());
    if (clampSide == HoverMode::kStart)
    {
@@ -144,19 +144,19 @@ void CanvasTimeline::MouseReleased()
 {
    if (mClick)
    {
-      float loopLength = mCanvas->mLoopEnd - mCanvas->mLoopStart;
+      double loopLength = mCanvas->mLoopEnd - mCanvas->mLoopStart;
       if (mHoverMode == HoverMode::kStart || mHoverMode == HoverMode::kMiddle)
       {
-         float startX = ofMap(mCanvas->mLoopStart, mCanvas->mViewStart, mCanvas->mViewEnd, 0, mWidth);
+         double startX = ofMap(mCanvas->mLoopStart, mCanvas->mViewStart, mCanvas->mViewEnd, 0, mWidth);
          startX += mDragOffset.x;
-         float quantized = GetQuantizedForX(startX, mHoverMode);
+         double quantized = GetQuantizedForX(startX, mHoverMode);
          mCanvas->mLoopStart = quantized;
       }
       if (mHoverMode == HoverMode::kEnd || mHoverMode == HoverMode::kMiddle)
       {
-         float endX = ofMap(mCanvas->mLoopEnd, mCanvas->mViewStart, mCanvas->mViewEnd, 0, mWidth);
+         double endX = ofMap(mCanvas->mLoopEnd, mCanvas->mViewStart, mCanvas->mViewEnd, 0, mWidth);
          endX += mDragOffset.x;
-         float quantized = GetQuantizedForX(endX, mHoverMode);
+         double quantized = GetQuantizedForX(endX, mHoverMode);
          mCanvas->mLoopEnd = quantized;
       }
