+++ Source/INoteReceiver.h
@@ -67,7 +67,7 @@ struct NoteMessage
 
    double time{ 0 };
    int pitch{ 0 };
-   float velocity{ 0 };
+   double velocity{ 0 };
    int voiceIdx{ -1 };
    ModulationParameters modulation;
 };
