+++ Source/LocationZoomer.cpp
@@ -49,7 +49,7 @@ void LocationZoomer::Update()
    if (mCurrentProgress < 1)
    {
       mCurrentProgress = ofClamp(mCurrentProgress + ofGetLastFrameTime() * mSpeed, 0, 1);
-      float ease;
+      double ease;
       if (mInVanityPanningMode) //ease in/out
          ease = mCurrentProgress < 0.5 ? 2 * mCurrentProgress * mCurrentProgress : 1 - pow(-2 * mCurrentProgress + 2, 2) / 2;
       else //ease out
