+++ Source/NoteChance.cpp
@@ -106,7 +106,7 @@ void NoteChance::DrawModule()
       ofSetColor(0, 255, 0);
       ofFill();
       double pos = fmod(TheTransport->GetMeasureTime(gTime) * TheTransport->GetTimeSigTop() / mLength, 1);
-      const float kPipSize = 3;
+      const double kPipSize = 3;
       float moduleWidth, moduleHeight;
       GetModuleDimensions(moduleWidth, moduleHeight);
       ofRect(ofMap(pos, 0, 1, 0, moduleWidth - kPipSize), lengthRect.y - 5, kPipSize, kPipSize);
