+++ Source/NoteFilter.h
@@ -55,7 +55,7 @@ private:
    void GetModuleDimensions(float& width, float& height) override;
 
    std::array<bool, 128> mGate;
-   std::array<float, 128> mLastPlayTime;
+   std::array<double, 128> mLastPlayTime;
    std::vector<Checkbox*> mGateCheckboxes;
    int mMinPitch{ 0 };
    int mMaxPitch{ 7 };
